+++ src/Gateway.php
@@ -11,6 +11,7 @@ use Pine\SimplePay\Payloads\RefundPayload;
 use Pine\SimplePay\Support\Config;
 use Pine\SimplePay\Support\Hash;
 use Pine\SimplePay\Support\Request;
+use Pine\SimplePay\Support\Str;
 use WC_Order;
 use WC_Payment_Gateway;
 
@@ -151,7 +152,7 @@ class Gateway extends WC_Payment_Gateway
     public function handlePayment()
     {
         $payload = json_decode(base64_decode($_GET['r']), true);
-        $order = wc_get_order(wc_get_order_id_by_order_key(strtolower($payload['o'])));
+        $order = wc_get_order(Str::idFromRef(strtolower($payload['o'])));
 
         if (! $order instanceof WC_Order) {
             wp_safe_redirect(wc_get_checkout_url());
@@ -172,7 +173,7 @@ class Gateway extends WC_Payment_Gateway
     {
         $input = file_get_contents('php://input');
         $payload = json_decode($input, true);
-        $order = wc_get_order(wc_get_order_id_by_order_key($payload['orderRef']));
+        $order = wc_get_order(Str::idFromRef($payload['orderRef']));
 
         if (! $order instanceof WC_Order) {
             die(__('Order not found.', 'pine-simplepay'));