+++ src/Plugin.php
@@ -12,7 +12,7 @@ class Plugin
      *
      * @var string
      */
-    public const VERSION = '2.3.1';
+    public const VERSION = '2.3.2';
 
     /**
      * The plugin slug.